requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.5
cycler==0.10.0
cymem==2.0.5
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
extract-msg==0.23.1
idna==2.10
IMAPClient==2.1.0
importlib-metadata==3.4.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
murmurhash==1.0.5
nltk==3.5
numpy==1.20.1
olefile==0.46
packaging==20.9
pandas==1.2.2
pathy==0.4.0
pdfminer.six==20201018
Pillow==8.1.0
preshed==3.0.5
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.12.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.2
spacy==3.0.3
spacy-legacy==3.0.1
SpeechRecognition==3.8.1
srsly==2.4.0
thinc==8.0.1
tqdm==4.56.2
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.3
wasabi==0.8.2
xlrd==1.2.0
XlsxWriter==1.3.7
zipp==3.4.0
